Cleanup more use of the "long" data type. This datatype varies in length between platforms so its use is not portable.

+++ redalert/bullet.cpp
@@ -533,7 +533,7 @@ int BulletClass::Shape_Number(void) const
     **	For tumbling projectiles, fetch offset stage.
     */
     if (Class->Tumble > 0) {
-        shapenum += (long)Frame % Class->Tumble;
+        shapenum += Frame % Class->Tumble;
     }
 
     return (shapenum);
